+++ app_feup/lib/utils/constants.dart
@@ -2,12 +2,25 @@ library Constants;
 
 const navPersonalArea = 'Área Pessoal';
 const navSchedule = 'Horário';
-const navExams = 'Mapa de Exames';
+const navExams = 'Exames';
 const navStops = 'Autocarros';
 const navAbout = 'Sobre';
 const navBugReport = 'Bugs e Sugestões';
 const navLogOut = 'Terminar sessão';
 
-const faculties = ['faup', 'fbaup', 'fcup', 'fcnaup', 'fadeup',
-                  'fdup', 'fep', 'feup', 'ffup', 'flup', 'fmup',
-                  'fmdup', 'fpceup', 'icbas'];
+const faculties = [
+  'faup',
+  'fbaup',
+  'fcup',
+  'fcnaup',
+  'fadeup',
+  'fdup',
+  'fep',
+  'feup',
+  'ffup',
+  'flup',
+  'fmup',
+  'fmdup',
+  'fpceup',
+  'icbas'
+];