+++ app_feup/lib/view/Pages/login_page_view.dart
@@ -71,7 +71,7 @@ class _LoginPageViewState extends State<LoginPageView> {
 
     return Scaffold(
         backgroundColor: Theme.of(context).brightness == Brightness.light
-            ? Theme.of(context).accentColor
+            ? Theme.of(context).primaryColor
             : Theme.of(context).scaffoldBackgroundColor,
         body: WillPopScope(
             child: Padding(
@@ -229,6 +229,12 @@ class _LoginPageViewState extends State<LoginPageView> {
   /// Creates the widget for the user to keep signed in (save his data).
   Widget createSaveDataCheckBox() {
     return CheckboxListTile(
+      activeColor: Theme.of(context).brightness == Brightness.light
+          ? Colors.white
+          : Colors.black54,
+      checkColor: Theme.of(context).brightness == Brightness.light
+          ? Theme.of(context).primaryColor
+          : Colors.white,
       value: _keepSignedIn,
       onChanged: _setKeepSignedIn,
       title: Text(
@@ -264,7 +270,7 @@ class _LoginPageViewState extends State<LoginPageView> {
           },
           child: Text('Entrar',
               style: TextStyle(
-                  color: Theme.of(context).accentColor,
+                  color: Theme.of(context).primaryColor,
                   fontWeight: FontWeight.w400,
                   fontSize: 20),
               textAlign: TextAlign.center),