+++ app_feup/lib/view/Widgets/navigation_drawer.dart
@@ -62,7 +62,7 @@ class NavigationDrawerState extends State<NavigationDrawer> {
         ? BoxDecoration(
             border: Border(
                 left: BorderSide(
-                    color: Theme.of(context).accentColor, width: 3.0)),
+                    color: Theme.of(context).primaryColor, width: 3.0)),
             color: Theme.of(context).dividerColor,
           )
         : null;
@@ -81,7 +81,7 @@ class NavigationDrawerState extends State<NavigationDrawer> {
             style: Theme.of(context)
                 .textTheme
                 .headline6
-                .apply(color: Theme.of(context).accentColor)),
+                .apply(color: Theme.of(context).primaryColor)),
       ),
     );
   }
@@ -112,7 +112,7 @@ class NavigationDrawerState extends State<NavigationDrawer> {
             child: Text(d,
                 style: TextStyle(
                     fontSize: 18.0,
-                    color: Theme.of(context).accentColor,
+                    color: Theme.of(context).primaryColor,
                     fontWeight: FontWeight.normal)),
           ),
           dense: true,