+++ app_feup/lib/view/Widgets/title_card.dart
@@ -14,16 +14,13 @@ class TitleCard extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return  Container(
+    return Container(
       margin: EdgeInsets.fromLTRB(12, 12, 12, 0),
       padding: EdgeInsets.only(top: 3, bottom: 3),
       alignment: Alignment.center,
-      child:  Text(
+      child: Text(
         '${this.weekDay}, ${this.day} de ${this.month}',
-        style: Theme.of(context)
-            .textTheme
-            .headline6
-            .apply(fontSizeDelta: 3, fontWeightDelta: -1),
+        style: Theme.of(context).textTheme.headline6,
       ),
     );
   }